[BE/FIX/#15] 아이템 - 매장 조회 변경

+++ Backend/src/main/java/com/hexa/muinus/common/exception/StoreLocationDuplicateException.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import java.math.BigDecimal;
 
 public class StoreLocationDuplicateException extends MuinusException {
-    public StoreLocationDuplicateException(BigDecimal x, BigDecimal y) {
+    public StoreLocationDuplicateException(Double x, Double y) {
         super("이미 등록된 위치입니다. locationX:" + x + ", locationY:" + y, 400);
     }
 }
